debug

+++ recovlm/models/qwen_2_5_vl/processing_qwen2_5_vl.py
@@ -308,7 +308,7 @@ class Qwen2_5_VLProcessor_moonvit(ProcessorMixin):
         )
         if images is not None:
             image_inputs = self.image_processor(images=images, return_tensors="pt")
-            image_inputs['pixel_values'] = image_inputs['pixel_values'].to(torch.bfloat16)
+            image_inputs['pixel_values'] = image_inputs['pixel_values']
             image_grid_thw = image_inputs["image_grid_thw"]
 
         else:
